update (v0.1) - complete redesign, added favicon, added react-hook-form for almost all forms on the site, added tabs in the profile (main, order history, admin), rewritten errors on the backend

+++ server/controllers/brandController.ts
@@ -1,45 +1,75 @@
 const { PrismaClient } = require('@prisma/client');
 const prisma = new PrismaClient();
+const ApiError = require('../error/apiError');
 
 class BrandController {
-    async create(req: any, res: any) {
-        const {name} = req.body;
-        const brand = await prisma.brand.create({
-            data: {name}
-        })
-        return res.json(brand);
-    }
-
-    async getAll(req: any, res: any) {
+    async create(req: any, res: any, next: any) {
         try {
-            const brands = await prisma.brand.findMany()
-            return res.json(brands);
+            const { name } = req.body;
+            if (!name) {
+                return next(ApiError.badRequest('Название бренда не передано.'))
+            }
+            const isExisting = await prisma.brand.findUnique({ where: { name } });
+
+            if (isExisting) {
+                return next(ApiError.badRequest('Данный бренд уже существует.'))
+            }
+
+            const brand = await prisma.brand.create({
+                data: { name }
+            })
+
+            return res.json(brand);
         } catch (err) {
-            return res.status(404).json({message: "Бренды не найдены"})
+            return next(ApiError.internal('Произошла ошибка на сервере. Попробуйте позже.'))
         }
     }
 
-    async getOne(req: any, res: any) {
-        const {id} = req.params;
-        const brand = await prisma.brand.findUnique({
-            where: {
-                id
-            }
-        })
-        return res.json(brand);
-    }
-    
-    async deleteOne(req: any, res: any) {
-        const {id} = req.params;
+    async getAll(req: any, res: any, next: any) {
         try {
+            const brands = await prisma.brand.findMany()
+            if (!brands) {
+                return next(ApiError.notFound('Не удалось получить список брендов.'))
+            }
+            return res.json(brands);
+        } catch (err) {
+            return next(ApiError.internal('Произошла ошибка на сервере. Попробуйте позже.'))
+        }
+    }
+
+    async getOne(req: any, res: any, next: any) {
+        try {
+            const { id } = req.params;
+            if (!id) {
+                return next(ApiError.notFound('Не найден ID бренда'))
+            }
+            const brand = await prisma.brand.findUnique({
+                where: {
+                    id
+                }
+            })
+            if (!brand) {
+                return next(ApiError.notFound('Бренд не найден'))
+            }
+            return res.json(brand);
+        } catch (err) {
+            return next(ApiError.internal('Произошла ошибка на сервере. Попробуйте позже.'))
+        }
+    }
+
+    async deleteOne(req: any, res: any, next: any) {
+        try {
+            const { id } = req.params;
+            if (!id) {
+                return next(ApiError.notFound('Не найден ID бренда'))
+            }
             const deleted = await prisma.brand.delete({
-                where: {id}
+                where: { id }
             })
 
-            res.json({message: "Бренд удален", deleted})
+            return res.json({ message: "Бренд удален", deleted })
         } catch (err) {
-            console.error(err);
-            res.status(500).json({error: "Ошибка сервера"})
+            return next(ApiError.internal('Произошла ошибка на сервере. Попробуйте позже.'))
         }
     }
 }
